move ii app to different file
